Fixed shaped crafting, added cheaper chains
Made chains give 3 instead of 1 when crafted, making chainmail a bit cheaper to craft. This is due to chainmail having the same protection rating as Gold, but having the same durability as Iron. Before, it was pointless to craft chainmail as it was weaker than iron yet cost a little bit more than iron armor due to the use of iron nuggets along with one iron ingot. Also, fixed crafting for the helmet and boots.